+++ CHANGELOG.rst
@@ -229,6 +229,8 @@ Added
   `#907 <https://github.com/openego/eGon-data/issues/907>`_
 * Add wind off shore power plants for eGon100RE
   `#868 <https://github.com/openego/eGon-data/issues/868>`_
+* Write simBEV metadata to DB table
+  `PR #978 <https://github.com/openego/eGon-data/pull/978>`_
 * Add desaggregation of pv home batteries onto buildings
   `#988 <https://github.com/openego/eGon-data/issues/988>`_